+++ tests/test_extensions/test_ext_autosummary.py
@@ -826,9 +826,8 @@ def test_autosummary_module_all(app):
         app.build()
         # generated/foo is generated successfully
         assert app.env.get_doctree('generated/autosummary_dummy_package_all')
-        module = (
-            app.srcdir / 'generated' / 'autosummary_dummy_package_all.rst'
-        ).read_text(encoding='utf8')
+        path = app.srcdir / 'generated' / 'autosummary_dummy_package_all.rst'
+        module = path.read_text(encoding='utf8')
         assert '   .. autosummary::\n   \n      PublicBar\n   \n' in module
         assert (
             '   .. autosummary::\n   \n      public_foo\n      public_baz\n   \n'
@@ -846,24 +845,20 @@ def test_autosummary_module_empty_all(app):
         app.build()
         # generated/foo is generated successfully
         assert app.env.get_doctree('generated/autosummary_dummy_package_empty_all')
-        module = (
-            app.srcdir / 'generated' / 'autosummary_dummy_package_empty_all.rst'
-        ).read_text(encoding='utf8')
+        path = app.srcdir / 'generated' / 'autosummary_dummy_package_empty_all.rst'
+        module = path.read_text(encoding='utf8')
         assert '.. automodule:: autosummary_dummy_package_empty_all' in module
-        # for __all__ = [], the output should not contain any variables
-        for name in [
-            '__all__',
-            '__builtins__',
-            '__cached__',
-            '__doc__',
-            '__file__',
-            '__loader__',
-            '__name__',
-            '__package__',
-            '__path__',
-            '__spec__',
-        ]:
-            assert name not in module
+        # for __all__ = (), the output should not contain any variables
+        assert '__all__' not in module
+        assert '__builtins__' not in module
+        assert '__cached__' not in module
+        assert '__doc__' not in module
+        assert '__file__' not in module
+        assert '__loader__' not in module
+        assert '__name__' not in module
+        assert '__package__' not in module
+        assert '__path__' not in module
+        assert '__spec__' not in module
     finally:
         sys.modules.pop('autosummary_dummy_package_all', None)